requirements.txt
numpy==1.21.5
pandas==1.3.5
streamlit==1.11.0
pyspark==3.3.1
kafka-python==2.0.2
scikit-learn==1.0.2